fix: 将 RoleOutputContract 定义移至 config/role_policy.py，消除循环导入
roles/definitions.py → execution/contracts.py → execution/__init__.py →
codeagent_runner.py → config/role_policy.py → roles/definitions.py 形成
循环导入，导致 test_module_imports_independently 失败。

修复：将 RoleOutputContract dataclass 定义移到 config/role_policy.py
（该模块只依赖 stdlib），roles/definitions.py 改为从 config/role_policy.py
导入，彻底断开循环链。
